Moved pages into seperate folders and fixed module paths
