+++ cob-fianance-app/src/App.js
@@ -1,18 +1,21 @@
 import React, {Component} from 'react';
 import {BrowserRouter as Router ,Switch, Route} from 'react-router-dom'
 
-import Dashboard from './components/pages/Dashboard';
-import DashboardI from './components/pages/DashboardI';
+
 import InstructorNavbar from './components/Layout/InstructorNavbar';
+import DashboardI from './components/pages/instructor/DashboardI';
+import UserInstructor from './components/pages/instructor/UserInstructor';
+import SettingsInstructor from './components/pages/instructor/SettingsInstuctor';
+import UserManagement from './components/pages/instructor/UserManagement';
+import SnapshotGroup from './components/pages/instructor/SnapshotGroup.js';
+
 import StudentNavbar from './components/Layout/StudentNavbar';
-import UserInstructor from './components/pages/UserInstructor';
-import SettingsInstructor from './components/pages/SettingsInstuctor';
-import UserManagement from './components/pages/UserManagement';
-import SnapshotGroup from './components/pages/SnapshotGroup.js';
-import Expenses from './components/pages/Expenses';
-import Transactions from './components/pages/Transactions';
+import Dashboard from './components/pages/student/Dashboard';
+import Expenses from './components/pages/student/Expenses';
+import Transactions from './components/pages/student/Transactions';
+import Page_Settings from './components/pages/student/Settings';
+
 import Login from './components/pages/Login';
-import Page_Settings from './components/pages/Settings';
 import Page404 from './components/pages/404';
 
 import './App.css';
