+++ cob-fianance-app/src/components/pages/instructor/DashboardI.js
@@ -1,6 +1,6 @@
 import React from 'react'
-import BusinessTable from '../Layout/BusinessTable';
-import AddBusinessDialogButton from '../Layout/AddBusinessDialogButton';
+import BusinessTable from '../../Layout/BusinessTable';
+import AddBusinessDialogButton from '../../Layout/AddBusinessDialogButton';
 
 export default function DashboardI() {
 	return (
