+++ cob-fianance-app/src/components/pages/student/Expenses.js
@@ -1,6 +1,6 @@
 import React from 'react';
-import ExpenseTable from '../Layout/ExpenseTable';
-import AddExpenseDialogButton from '../Layout/AddExpenseDialogButton';
+import ExpenseTable from '../../Layout/ExpenseTable';
+import AddExpenseDialogButton from '../../Layout/AddExpenseDialogButton';
 
 export default function Expenses() {
 	return (
@@ -10,7 +10,7 @@ export default function Expenses() {
 			<div style={{textAlign: 'right', margin: '5px'}}>
 				<AddExpenseDialogButton/>
 			</div>
-			<ExpenseTable style = {{paddingTop: '10px 20px' }}></ExpenseTable>
+				<ExpenseTable style = {{paddingTop: '10px 20px' }}></ExpenseTable>
 			</div>
 		</>
 	)
