+++ cob-fianance-app/src/components/pages/student/Settings.js
@@ -1,5 +1,5 @@
 import React from 'react'
-import ProfitGoalsForm from '../Layout/ProfitGoalsForm';
+import ProfitGoalsForm from '../../Layout/ProfitGoalsForm';
 
 export default function Settings() {
 	return (
